# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NovaPay fintech API
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export interface MessageResponse {
  message: string;
}

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName: string;
  lastName: string;
  phone?: string;
}

export interface LoginRequest {
  email: string;
  password: string;
}

export interface ForgotPasswordRequest {
  email: string;
}

export type UserProfileKycStatus =
  (typeof UserProfileKycStatus)[keyof typeof UserProfileKycStatus];

export const UserProfileKycStatus = {
  pending: "pending",
  submitted: "submitted",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface UserProfile {
  id: string;
  email: string;
  firstName: string;
  lastName: string;
  phone?: string | null;
  avatarUrl?: string | null;
  kycStatus: UserProfileKycStatus;
  createdAt: string;
}

export interface AuthResponse {
  token: string;
  user: UserProfile;
}

export interface UpdateProfileRequest {
  firstName?: string;
  lastName?: string;
  phone?: string;
}

export interface WalletBalance {
  balance: number;
  currency: string;
  accountNumber: string;
}

export interface BalanceUpdate {
  balance: number;
  currency: string;
}

export interface MoneyRequest {
  /** @minimum 0.01 */
  amount: number;
  note?: string;
}

export type VirtualCardCardType =
  (typeof VirtualCardCardType)[keyof typeof VirtualCardCardType];

export const VirtualCardCardType = {
  visa: "visa",
  mastercard: "mastercard",
} as const;

export interface VirtualCard {
  id?: string;
  cardNumber: string;
  cardHolder: string;
  expiryMonth: number;
  expiryYear: number;
  cvv: string;
  cardType: VirtualCardCardType;
  isActive: boolean;
}

export interface SendMoneyRequest {
  recipientEmail: string;
  /** @minimum 0.01 */
  amount: number;
  note?: string;
}

export type TransactionRecordType =
  (typeof TransactionRecordType)[keyof typeof TransactionRecordType];

export const TransactionRecordType = {
  credit: "credit",
  debit: "debit",
} as const;

export type TransactionRecordStatus =
  (typeof TransactionRecordStatus)[keyof typeof TransactionRecordStatus];

export const TransactionRecordStatus = {
  pending: "pending",
  completed: "completed",
  failed: "failed",
} as const;

export interface TransactionRecord {
  id: string;
  type: TransactionRecordType;
  amount: number;
  currency: string;
  description: string;
  status: TransactionRecordStatus;
  counterpartyName?: string | null;
  counterpartyEmail?: string | null;
  createdAt: string;
}

export interface TransactionList {
  transactions: TransactionRecord[];
  total: number;
  page: number;
  limit: number;
  hasMore: boolean;
}

export type KycStatusStatus =
  (typeof KycStatusStatus)[keyof typeof KycStatusStatus];

export const KycStatusStatus = {
  pending: "pending",
  submitted: "submitted",
  approved: "approved",
  rejected: "rejected",
} as const;

export interface KycStatus {
  status: KycStatusStatus;
  submittedAt?: string | null;
  reviewedAt?: string | null;
  rejectionReason?: string | null;
}

export type KycSubmitRequestDocumentType =
  (typeof KycSubmitRequestDocumentType)[keyof typeof KycSubmitRequestDocumentType];

export const KycSubmitRequestDocumentType = {
  passport: "passport",
  national_id: "national_id",
  drivers_license: "drivers_license",
} as const;

export interface KycSubmitRequest {
  fullName: string;
  dateOfBirth: string;
  nationality: string;
  documentType: KycSubmitRequestDocumentType;
  documentNumber: string;
  addressLine1: string;
  city: string;
  country: string;
}

export type GetTransactionsParams = {
  page?: number;
  limit?: number;
};