# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.2.0
  description: NovaPay fintech API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: auth
    description: Authentication operations
  - name: user
    description: User profile operations
  - name: wallet
    description: Wallet operations
  - name: transactions
    description: Transaction operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /auth/register:
    post:
      operationId: authRegister
      tags: [auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
      responses:
        "201":
          description: User registered successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Email already exists
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/login:
    post:
      operationId: authLogin
      tags: [auth]
      summary: Login user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /auth/forgot-password:
    post:
      operationId: authForgotPassword
      tags: [auth]
      summary: Request password reset
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "200":
          description: Reset link sent (if email exists)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /user/profile:
    get:
      operationId: getUserProfile
      tags: [user]
      summary: Get current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    patch:
      operationId: updateUserProfile
      tags: [user]
      summary: Update user profile
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Profile updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallet/balance:
    get:
      operationId: getWalletBalance
      tags: [wallet]
      summary: Get wallet balance
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Wallet balance info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WalletBalance"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallet/card:
    get:
      operationId: getVirtualCard
      tags: [wallet]
      summary: Get virtual card details
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Virtual card info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualCard"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallet/card/freeze:
    post:
      operationId: freezeCard
      tags: [wallet]
      summary: Freeze virtual card
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Card frozen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualCard"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallet/card/unfreeze:
    post:
      operationId: unfreezeCard
      tags: [wallet]
      summary: Unfreeze virtual card
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Card unfrozen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VirtualCard"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Card not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallet/add:
    post:
      operationId: addMoney
      tags: [wallet]
      summary: Add money to wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoneyRequest"
      responses:
        "200":
          description: Money added successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceUpdate"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallet/withdraw:
    post:
      operationId: withdrawMoney
      tags: [wallet]
      summary: Withdraw money from wallet
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MoneyRequest"
      responses:
        "200":
          description: Withdrawal successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BalanceUpdate"
        "400":
          description: Insufficient funds or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /wallet/send:
    post:
      operationId: sendMoney
      tags: [wallet]
      summary: Send money to another user
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendMoneyRequest"
      responses:
        "200":
          description: Transfer successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionRecord"
        "400":
          description: Insufficient funds or validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /transactions:
    get:
      operationId: getTransactions
      tags: [transactions]
      summary: Get transaction history
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Transaction list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /kyc/status:
    get:
      operationId: getKycStatus
      tags: [user]
      summary: Get KYC verification status
      security:
        - bearerAuth: []
      responses:
        "200":
          description: KYC status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /kyc/submit:
    post:
      operationId: submitKyc
      tags: [user]
      summary: Submit KYC verification
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/KycSubmitRequest"
      responses:
        "200":
          description: KYC submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/KycStatus"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    RegisterRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
          minLength: 8
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
      required:
        - email
        - password
        - firstName
        - lastName

    LoginRequest:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required:
        - email
        - password

    ForgotPasswordRequest:
      type: object
      properties:
        email:
          type: string
          format: email
      required:
        - email

    AuthResponse:
      type: object
      properties:
        token:
          type: string
        user:
          $ref: "#/components/schemas/UserProfile"
      required:
        - token
        - user

    UserProfile:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string
          nullable: true
        avatarUrl:
          type: string
          nullable: true
        kycStatus:
          type: string
          enum: [pending, submitted, approved, rejected]
        createdAt:
          type: string
      required:
        - id
        - email
        - firstName
        - lastName
        - kycStatus
        - createdAt

    UpdateProfileRequest:
      type: object
      properties:
        firstName:
          type: string
        lastName:
          type: string
        phone:
          type: string

    WalletBalance:
      type: object
      properties:
        balance:
          type: number
        currency:
          type: string
        accountNumber:
          type: string
      required:
        - balance
        - currency
        - accountNumber

    BalanceUpdate:
      type: object
      properties:
        balance:
          type: number
        currency:
          type: string
      required:
        - balance
        - currency

    MoneyRequest:
      type: object
      properties:
        amount:
          type: number
          minimum: 0.01
        note:
          type: string
      required:
        - amount

    VirtualCard:
      type: object
      properties:
        id:
          type: string
        cardNumber:
          type: string
        cardHolder:
          type: string
        expiryMonth:
          type: integer
        expiryYear:
          type: integer
        cvv:
          type: string
        cardType:
          type: string
          enum: [visa, mastercard]
        isActive:
          type: boolean
      required:
        - cardNumber
        - cardHolder
        - expiryMonth
        - expiryYear
        - cvv
        - cardType
        - isActive

    SendMoneyRequest:
      type: object
      properties:
        recipientEmail:
          type: string
          format: email
        amount:
          type: number
          minimum: 0.01
        note:
          type: string
      required:
        - recipientEmail
        - amount

    TransactionRecord:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [credit, debit]
        amount:
          type: number
        currency:
          type: string
        description:
          type: string
        status:
          type: string
          enum: [pending, completed, failed]
        counterpartyName:
          type: string
          nullable: true
        counterpartyEmail:
          type: string
          nullable: true
        createdAt:
          type: string
      required:
        - id
        - type
        - amount
        - currency
        - description
        - status
        - createdAt

    TransactionList:
      type: object
      properties:
        transactions:
          type: array
          items:
            $ref: "#/components/schemas/TransactionRecord"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
        hasMore:
          type: boolean
      required:
        - transactions
        - total
        - page
        - limit
        - hasMore

    KycStatus:
      type: object
      properties:
        status:
          type: string
          enum: [pending, submitted, approved, rejected]
        submittedAt:
          type: string
          nullable: true
        reviewedAt:
          type: string
          nullable: true
        rejectionReason:
          type: string
          nullable: true
      required:
        - status

    KycSubmitRequest:
      type: object
      properties:
        fullName:
          type: string
        dateOfBirth:
          type: string
        nationality:
          type: string
        documentType:
          type: string
          enum: [passport, national_id, drivers_license]
        documentNumber:
          type: string
        addressLine1:
          type: string
        city:
          type: string
        country:
          type: string
      required:
        - fullName
        - dateOfBirth
        - nationality
        - documentType
        - documentNumber
        - addressLine1
        - city
        - country