# lib/api-zod/src/generated/types/kycSubmitRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NovaPay fintech API
 * OpenAPI spec version: 0.2.0
 */
import type { KycSubmitRequestDocumentType } from "./kycSubmitRequestDocumentType";

export interface KycSubmitRequest {
  fullName: string;
  dateOfBirth: string;
  nationality: string;
  documentType: KycSubmitRequestDocumentType;
  documentNumber: string;
  addressLine1: string;
  city: string;
  country: string;
}