# lib/api-zod/src/generated/types/loginRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NovaPay fintech API
 * OpenAPI spec version: 0.2.0
 */

export interface LoginRequest {
  email: string;
  password: string;
}