# lib/api-zod/src/generated/types/registerRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NovaPay fintech API
 * OpenAPI spec version: 0.2.0
 */

export interface RegisterRequest {
  email: string;
  /** @minLength 8 */
  password: string;
  firstName: string;
  lastName: string;
  phone?: string;
}