# lib/api-zod/src/generated/types/sendMoneyRequest.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NovaPay fintech API
 * OpenAPI spec version: 0.2.0
 */

export interface SendMoneyRequest {
  recipientEmail: string;
  /** @minimum 0.01 */
  amount: number;
  note?: string;
}