# lib/api-zod/src/generated/types/transactionRecord.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NovaPay fintech API
 * OpenAPI spec version: 0.2.0
 */
import type { TransactionRecordStatus } from "./transactionRecordStatus";
import type { TransactionRecordType } from "./transactionRecordType";

export interface TransactionRecord {
  id: string;
  type: TransactionRecordType;
  amount: number;
  currency: string;
  description: string;
  status: TransactionRecordStatus;
  counterpartyName?: string | null;
  counterpartyEmail?: string | null;
  createdAt: string;
}