# lib/api-zod/src/generated/types/walletBalance.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * NovaPay fintech API
 * OpenAPI spec version: 0.2.0
 */

export interface WalletBalance {
  balance: number;
  currency: string;
  accountNumber: string;
}